pyproject.toml
[project]
name = "nonebot-plugin-gold-price"
version = "0.1.8"
description = "用于查询实时金价、绘制历史金价的NoneBot2插件"
authors = [
    { name = "newcovid", email = "280310454@qq.com" },
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "GPL-3.0" }
keywords = ["nonebot", "gold", "finance", "analysis"]
dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-adapter-onebot>=2.3.0",
    "aiohttp >=3.8.0",
    "matplotlib>=3.6.0",
    "nonebot-plugin-apscheduler>=0.3.0",
    "nonebot-plugin-localstore>=0.7.0",
    "pydantic>=1.10.0"
]

[project.urls]
Homepage = "https://github.com/newcovid/nonebot-plugin-gold-price"
Documentation = "https://github.com/newcovid/nonebot-plugin-gold-price#readme"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["nonebot_plugin_gold_price"]